libretro-common: Update zlib headers to 1.2.11.
Updated from a modified version 1.2.8, April 28th, 2013.
