+++ deps/libz/gzguts.h
@@ -19,7 +19,7 @@
 #endif
 
 #include <stdio.h>
-#include "zlib.h"
+#include <zlib.h>
 #ifdef STDC
 #  include <string.h>
 #  include <stdlib.h>
