+++ deps/libz/zutil.h
@@ -19,7 +19,7 @@
 #  define ZLIB_INTERNAL
 #endif
 
-#include "zlib.h"
+#include <zlib.h>
 
 #if defined(STDC) && !defined(Z_SOLO)
 #  if !(defined(_WIN32_WCE) && defined(_MSC_VER))
